Move PatchCore's coreset helpers to vision utilities
RescaleSegmentor and ApproximateGreedyCoresetSampler are needed verbatim by
every model that scores patches against a subsampled memory bank, so they move
out of the PatchCore package into pyclad.vision.models.utilities.coreset. UCAD,
on the vision_ucad branch, needs exactly these two and would otherwise carry a
second copy; importing them from PatchCore instead would make one model depend
on another's package.

The moved sampler gains an exact-count entry point (run_with_target_size) next
to the percentage-based run(), and percentage becomes optional so a caller that
sizes its coreset absolutely can leave it unset. PatchCore's call sites are
keyword-only and unchanged; run() is bit-for-bit identical to the version
removed here, verified against it across pool sizes and percentages.

The file is byte-identical to the one on vision_ucad, so the two branches merge
without a conflict on it in either order.
