+++ src/pyclad/vision/models/patchcore/patchcore.py
@@ -3,7 +3,6 @@ from typing import Optional, Sequence
 import numpy as np
 import torch
 import torch.nn.functional as F
-from scipy import ndimage
 from sklearn.neighbors import NearestNeighbors
 from torch import nn
 
@@ -13,6 +12,10 @@ from pyclad.vision.models.utilities.backbones import (
     default_backbone_return_nodes,
 )
 from pyclad.vision.models.utilities.base_model import VisionScoringBase
+from pyclad.vision.models.utilities.coreset import (
+    ApproximateGreedyCoresetSampler,
+    RescaleSegmentor,
+)
 
 
 class MeanMapper(nn.Module):
@@ -80,105 +83,6 @@ class PatchMaker:
         while x_t.ndim > 1:
             x_t = torch.max(x_t, dim=-1).values
         return x_t.numpy() if isinstance(x, np.ndarray) else x_t
-
-
-class RescaleSegmentor:
-    def __init__(self, device: torch.device, target_size: tuple[int, int], smoothing: float):
-        self.device = device
-        self.target_size = target_size
-        self.smoothing = smoothing
-
-    def convert_to_segmentation(self, patch_scores: np.ndarray) -> np.ndarray:
-        with torch.no_grad():
-            if isinstance(patch_scores, np.ndarray):
-                patch_scores = torch.from_numpy(patch_scores)
-
-            scores = patch_scores.to(self.device).unsqueeze(1)
-            scores = F.interpolate(scores, size=self.target_size, mode="bilinear", align_corners=False)
-            scores = scores.squeeze(1).cpu().numpy()
-
-        if self.smoothing <= 0:
-            return scores.astype(np.float32, copy=False)
-        return np.asarray(
-            [ndimage.gaussian_filter(score, sigma=self.smoothing) for score in scores],
-            dtype=np.float32,
-        )
-
-
-class ApproximateGreedyCoresetSampler:
-    def __init__(
-        self,
-        percentage: float,
-        device: torch.device,
-        number_of_starting_points: int = 10,
-        dimension_to_project_features_to: int = 128,
-        random_seed: int = 0,
-    ):
-        if percentage <= 0.0 or percentage > 1.0:
-            raise ValueError("percentage must be in (0, 1]")
-
-        self.percentage = percentage
-        self.device = device
-        self.number_of_starting_points = number_of_starting_points
-        self.dimension_to_project_features_to = dimension_to_project_features_to
-        self.random_seed = random_seed
-
-    def _reduce_features(self, features: torch.Tensor) -> torch.Tensor:
-        if features.shape[1] == self.dimension_to_project_features_to:
-            return features.to(self.device)
-
-        with torch.random.fork_rng():
-            torch.manual_seed(self.random_seed)
-            mapper = nn.Linear(features.shape[1], self.dimension_to_project_features_to, bias=False).to(self.device)
-        return mapper(features.to(self.device))
-
-    @staticmethod
-    def _compute_batchwise_differences(matrix_a: torch.Tensor, matrix_b: torch.Tensor) -> torch.Tensor:
-        a_times_a = matrix_a.unsqueeze(1).bmm(matrix_a.unsqueeze(2)).reshape(-1, 1)
-        b_times_b = matrix_b.unsqueeze(1).bmm(matrix_b.unsqueeze(2)).reshape(1, -1)
-        a_times_b = matrix_a.mm(matrix_b.T)
-        return (-2 * a_times_b + a_times_a + b_times_b).clamp(0, None).sqrt()
-
-    def _compute_greedy_coreset_indices(self, features: torch.Tensor) -> np.ndarray:
-        """Greedy coreset selection: repeatedly pick the point with the largest approximate
-        min-distance to the points already selected.
-        """
-        number_of_starting_points = min(self.number_of_starting_points, len(features))
-        rng = np.random.default_rng(self.random_seed)
-        start_points = rng.choice(len(features), number_of_starting_points, replace=False).tolist()
-
-        approximate_distance_matrix = self._compute_batchwise_differences(features, features[start_points])
-        approximate_coreset_anchor_distances = torch.mean(approximate_distance_matrix, axis=-1).reshape(-1, 1)
-
-        coreset_indices = []
-        num_coreset_samples = max(1, int(len(features) * self.percentage))
-
-        with torch.no_grad():
-            for _ in range(num_coreset_samples):
-                select_idx = torch.argmax(approximate_coreset_anchor_distances.squeeze(-1)).item()
-                coreset_indices.append(select_idx)
-
-                coreset_select_distance = self._compute_batchwise_differences(
-                    features, features[select_idx : select_idx + 1]
-                )
-                approximate_coreset_anchor_distances = torch.cat(
-                    [approximate_coreset_anchor_distances, coreset_select_distance],
-                    dim=-1,
-                )
-                approximate_coreset_anchor_distances = torch.min(
-                    approximate_coreset_anchor_distances, dim=1
-                ).values.reshape(-1, 1)
-
-        return np.array(coreset_indices)
-
-    def run(self, features: np.ndarray) -> np.ndarray:
-        if self.percentage == 1.0:
-            return features
-
-        feature_tensor = torch.from_numpy(features.astype(np.float32, copy=False))
-        reduced_features = self._reduce_features(feature_tensor)
-        sample_indices = self._compute_greedy_coreset_indices(reduced_features)
-        return feature_tensor[sample_indices].cpu().numpy().astype(np.float32, copy=False)
 
 
 class PatchCore(VisionScoringBase):
